+++ build_runner/lib/src/bootstrap/build_process_state.dart
@@ -7,6 +7,10 @@ import 'dart:convert';
 import 'dart:io';
 import 'dart:isolate';
 
+import '../build_plan/build_paths.dart';
+
+import 'build_process_lock.dart';
+
 /// State for the whole build process.
 ///
 /// It's passed from the entrypoint to the spawned build script isolate, then
@@ -17,6 +21,45 @@ class BuildProcessState {
   final Map<String, Object?> _state = {};
   final List<void Function()> _beforeSends = [];
   final List<void Function()> _afterReceives = [];
+  BuildProcessLock? _lock;
+
+  BuildProcessState() {
+    _afterReceives.add(() {
+      // Initialize lock using state from parent process if it was set.
+      if (_buildPaths != null) {
+        _lock = BuildProcessLock(_buildPaths!);
+      }
+    });
+  }
+
+  /// Configures and takes the [BuildProcessLock].
+  ///
+  /// It will be released on process exit.
+  ///
+  /// If it can't be taken, notifies the lock holder and waits for it to become
+  /// available.
+  ///
+  /// Throws if called more than once.
+  Future<void> takeLock(BuildPaths buildPaths) {
+    if (_lock != null) throw StateError('Already took lock.');
+    _buildPaths = buildPaths;
+    _lock = BuildProcessLock(buildPaths);
+    return _lock!.takeLock();
+  }
+
+  /// Registers a callback to be called when a lock is requested.
+  ///
+  /// Setting the callback starts the watch for lock request files.
+  void setLockRequestCallback(void Function() callback) {
+    // _lock is null in unit tests, which never get lock requests.
+    _lock?.setLockRequestCallback(callback);
+  }
+
+  /// Whether another process has requested the [BuildProcessLock].
+  bool isLockRequested() {
+    // _lock is null in unit tests, which never get lock requests.
+    return _lock?.isLockRequested() ?? false;
+  }
 
   /// For `buildLog`, console output capabilities.
   ///
@@ -63,6 +106,21 @@ class BuildProcessState {
   String get packageConfigUri =>
       (_state['packageConfigUri'] ??= Isolate.packageConfigSync!.toString())
           as String;
+
+  /// The [BuildPaths] used for the process lock.
+  BuildPaths? get _buildPaths =>
+      _state.containsKey('packagePath')
+          ? BuildPaths(
+            packagePath: _state['packagePath'] as String,
+            workspacePath: _state['workspacePath'] as String?,
+            buildWorkspace: _state['buildWorkspace'] as bool,
+          )
+          : null;
+  set _buildPaths(BuildPaths? value) {
+    _state['packagePath'] = value?.packagePath;
+    _state['workspacePath'] = value?.workspacePath;
+    _state['buildWorkspace'] = value?.buildWorkspace;
+  }
 
   void resetForTests() {
     _state.clear();